Edits to indexing and style
Some changes to the blog page styling. Added a header element with class for
each post. Styled to include description, etc. The downside is that it
is causing a linebreak between the date and the body of the post that I
wasn't looking for. This happens on the index page too, between the
title and the description. They were only kept inline because of a
previous hacky solution.

Changed the incoming links from changes.md to match
the pattern for the URL. The current solution is quite hacky, and
depends on slugifying the title. I believe it's because of the custom
URL we have set in config.

Also styled second paragraph to indent.