Updates to blog page
Changed structure and style of blog page. Changed URLs of posts to their dates. Still need to fix the issue with blockquotes and other elements, so that they appear on their own line. Also need to add in templating in the event that a post does have a title. Need to change the alt text for links to a conditional. The style of this page is now based on how I have seen diaries be printed, and how I write my own: the date, a dash, and the content beginning on the same line. We also need to change the TOC so that it just lists months with a post count next to them (maybe), or to have collapsible sections. We should also still make use of the description metadata field, which is useful for the index page.