feat: new blog post, blog section sorting on index
Added a new post about the party I went to at Drones Club. Took a few
hours to write. Also noticed that the index page was only showing the
first ten posts in that collection: fixed it. I have decided that I
don't like the design of the blog section on the index page.