add/fix: diary and blog page
added some edited selections from the diary, covering september and
october 2023. also made several modifications to the blog page. for the
moment, it's necessary to have the date field in the blog collection
metadata in order for things to show up properly.

i'm still working on adding the ideal image slider plugin. also need to
remove those hideous scotch lines from the blog page, and fix the titles
for untitled entries. we want a way to support titles and descriptions,
	but in general, diary writing is not built around titled entries.