add/fix
added a lot of styling, primarily to headers on pages. this involved a
fairly significant re-write of the website, even requiring that i
activate a new pandoc plugin in order to wrap content blocks in
<section> tags. i experimented with adding hyperlinks to the section
text manually, but it broke the TOC (compare /links to /about). i also
tried to get automatic injection of anchor tags working through
injec_anchors.rb, but it only worked on some pages and not others. there
are ongoing issues so that there is now inconsistency between pages, but
that was a significant amount of progress.