add github links
semi-automatically-generated links to github commits on all diary
entries, as well as to the posts layout. the solution is hacky in that
it requires manual construction of part of the url (branch and blob). it
also only currently works for the newest revision date (i couldn't
figure out how to add the initial commit without cheating and doing it
manually).