add/fix: style, layout
moved sidebar from side to top. changed the layouts and media sizes to
still be mobile-first, but without artificial margins.

spent a lot of time trying to get anchors working for articles. ended up
removing the new article layout and re-writing the css to apply the
::after to all headings, but this means we still have the bug on the
index page. it also means that we still haven't figured out how to
inject anchors into article headings. if the links are manually added,
it still fails to properly format.