add/fix: style, diaries
a big commit that changes /blog to /diaries. changes the format of the
diary from a standard reverse-chrono blog page, to a chronological diary
page that is separated into numbered chapters. the chapters have been
edited for brevity, trying to find the arc, but there's still a lot of
repetition to remove. certain styling on the diary page has been changed
to distinguish it from the blog page, notably around the headers.

a big obstacle that remains, that i can't figure out how to resolve, is
adding links to the chapter sub-sections. ideally we would be able to
link to each entry, the same as with the blog page.

we are also still missing entries that i want to include in chapter III.
i would say the total arc will comprise a chapter IV, and will end
around mother's day. a new narrative arc begins with chapter V / seeing
kat.