edits to the text
did extensive edits to the text. a lot of work on the "harvest" story,
as well as a lot of work combining the "new media and literature" essay
into the "about" page. also moved the "commentary" page and added a link
to the index. other miscellaneous improvements all over the place. still a lot of work to go.