edits to newfoundland
absolutely discouraged today after only a few minutes of work. this
story has no conceptual gimmick, no imagery that makes it stand
out---but my fixation on having a gimmick is itself a problem. henry
james didn't have gimmicks, he didn't bring in radical imagery or
conceptual vocabulary. the closest thing to a "gimmick" he has are his
sentences, and the structures themselves. otherwise the story stands for
itself. this story is completely draining me, and it holds no interest.
no-one will ever want to publish it. writing to publication is folly.

this is what i have given my life to. sacrificed my happiness to. "can't
try to fix myself and be a writer at the same time." it's not worth it,
it's not worth it: no-one cares!! this story is not transcendent. it's a
fucking cliche. i have given up any hope of happiness for this bullshit.
it's not worth it: it's just a way to hide from my problems. it's easier
to continue suffering than to actually overcome your issues. it's easier
to hide than to confront. and i am hiding in this absolute dreck,
hackwork, dribble, maudlin bullshit. kill me...i'm not an artist, never
was, just another cowardly egotist.